ref: Make voting class more error resistant (#198)
* ref: Make voting class more error resistant

* style: format code with Autopep8 and isort

This commit fixes the style issues introduced in 0f7685c according to the output
from Autopep8 and isort.

Details: https://github.com/FGBxRamel/CommunityHelperBot/pull/198

---------

+++ classes/voting.py
@@ -1,10 +1,12 @@
-from random import randint
-from time import localtime, sleep, strftime, time
-import interactions as i
 import configparser as cp
-import classes.database as db
-from sqlite3 import IntegrityError
 import re
+from random import randint
+from sqlite3 import IntegrityError
+from time import localtime, sleep, strftime, time
+
+import interactions as i
+
+import classes.database as db
 
 
 class Voting():
@@ -83,12 +85,17 @@ class Voting():
         self._save()
         embed = await self._get_embed()
         message = await self.channel.fetch_message(self.message_id)
+        if message is None:
+            await self.delete()
+            return
         await message.edit(embed=embed)
 
     async def delete(self) -> None:
         """Deletes the voting from the database and the embed."""
         db.delete_data("votings", {"voting_id": self.id})
         message = await self.channel.fetch_message(self.message_id)
+        if message is None:
+            return
         try:
             await message.delete()
         except AttributeError:
@@ -96,6 +103,9 @@ class Voting():
 
     async def close(self) -> None:
         message = await self.channel.fetch_message(self.message_id)
+        if message is None:
+            await self.delete()
+            return
         try:
             tie = await self._is_tie()
         except AttributeError:
